Refactoring and documenting of the runtime dependency registry
