+++ sdk/src/Core/Amazon.Runtime/AmazonServiceClient.cs
@@ -17,7 +17,6 @@ using Amazon.Runtime.Internal;
 using Amazon.Runtime.Internal.Auth;
 using Amazon.Runtime.Internal.Transform;
 using Amazon.Runtime.Internal.Util;
-using Amazon.RuntimeDependencyRegistry;
 using Amazon.Util;
 using System;
 using System.Collections.Generic;
@@ -39,7 +38,6 @@ namespace Amazon.Runtime
 
         private bool _disposed;
         private Logger _logger;
-        public ServiceClientRuntimeDependencyRegistry RuntimeDependencyRegistry { get; } = new ServiceClientRuntimeDependencyRegistry();
 
         protected EndpointDiscoveryResolverBase EndpointDiscoveryResolver { get; private set; }
         protected RuntimePipeline RuntimePipeline { get; set; }
@@ -413,9 +411,6 @@ namespace Amazon.Runtime
             {
                 if (RuntimePipeline != null)
                     RuntimePipeline.Dispose();
-
-                if(RuntimeDependencyRegistry != null)
-                    RuntimeDependencyRegistry.Dispose();
 
                 _disposed = true;
             }
